// src/api/generated/model/responsibleDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Core | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UserDTO } from './userDTO.ts';

export interface ResponsibleDTO {
  operationId: string;
  userId: string;
  user: UserDTO;
  id: string;
  createdAt: string;
  updatedAt: string;
}